move side-effect handling to `PermissionController`
