+++ packages/approval-controller/src/ApprovalController.test.ts
@@ -591,138 +591,10 @@ describe('approval controller', () => {
       expect(deleteSpy.callCount).toStrictEqual(numDeletions);
     });
 
-    it('triggers permitted handlers', async () => {
-      numDeletions = 1;
-      const permitterHandlerMock = jest.fn(() => Promise.resolve('foo'));
-      const failureHandlerMock = jest.fn(() => Promise.resolve());
-
-      const approvalPromise = approvalController.add({
-        id: 'foo',
-        origin: 'bar.baz',
-        type: 'myType',
-        sideEffects: {
-          permittedHandlers: [permitterHandlerMock, permitterHandlerMock],
-          failureHandlers: [failureHandlerMock],
-        },
-      });
-      await approvalController.accept('foo', 'success');
-
-      const result = await approvalPromise;
-      expect(result).toStrictEqual({
-        sideEffectsData: ['foo', 'foo'],
-        value: 'success',
-      });
-      expect(permitterHandlerMock).toHaveBeenCalledTimes(2);
-      expect(failureHandlerMock).not.toHaveBeenCalled();
-      expect(deleteSpy.callCount).toStrictEqual(numDeletions);
-    });
-
-    it('triggers failure handlers on rejected permitted handler', async () => {
-      numDeletions = 1;
-      const errorMessage = 'failed';
-      const permitterHandlerMock = jest.fn(() =>
-        Promise.reject(new Error(errorMessage)),
+    it('throws on unknown id', () => {
+      expect(() => approvalController.accept('foo')).toThrow(
+        getIdNotFoundError('foo'),
       );
-      const failureHandlerMock = jest.fn(() => Promise.resolve());
-
-      const approvalPromise = approvalController.add({
-        id: 'foo',
-        origin: 'bar.baz',
-        type: 'myType',
-        sideEffects: {
-          permittedHandlers: [permitterHandlerMock],
-          failureHandlers: [failureHandlerMock, failureHandlerMock],
-        },
-      });
-      await approvalController.accept('foo', 'success');
-
-      await expect(async () => approvalPromise).rejects.toThrow(errorMessage);
-      expect(permitterHandlerMock).toHaveBeenCalled();
-      expect(failureHandlerMock).toHaveBeenCalledTimes(2);
-      expect(deleteSpy.callCount).toStrictEqual(numDeletions);
-    });
-
-    it('throws the appropriate error if there is no failure handler', async () => {
-      numDeletions = 1;
-      const errorMessage = 'failed';
-      const permitterHandlerMock = jest.fn(() =>
-        Promise.reject(new Error(errorMessage)),
-      );
-
-      const approvalPromise = approvalController.add({
-        id: 'foo',
-        origin: 'bar.baz',
-        type: 'myType',
-        sideEffects: {
-          permittedHandlers: [permitterHandlerMock],
-        },
-      });
-      await approvalController.accept('foo', 'success');
-
-      await expect(async () => approvalPromise).rejects.toThrow(errorMessage);
-      expect(permitterHandlerMock).toHaveBeenCalled();
-      expect(deleteSpy.callCount).toStrictEqual(numDeletions);
-    });
-
-    it('throws the appropriate error on multiple rejected permitted handlers', async () => {
-      numDeletions = 1;
-      const errorMessage = 'failed';
-      const permitterHandlerMock = jest.fn(() =>
-        Promise.reject(new Error(errorMessage)),
-      );
-      const failureHandlerMock = jest.fn(() => Promise.resolve());
-
-      const approvalPromise = approvalController.add({
-        id: 'foo',
-        origin: 'bar.baz',
-        type: 'myType',
-        sideEffects: {
-          permittedHandlers: [permitterHandlerMock, permitterHandlerMock],
-          failureHandlers: [failureHandlerMock, failureHandlerMock],
-        },
-      });
-      await approvalController.accept('foo', 'success');
-
-      await expect(async () => approvalPromise).rejects.toThrow(
-        'Multiple errors appened during side-effects execution',
-      );
-      expect(permitterHandlerMock).toHaveBeenCalled();
-      expect(failureHandlerMock).toHaveBeenCalledTimes(2);
-      expect(deleteSpy.callCount).toStrictEqual(numDeletions);
-    });
-
-    it('rejects the approval if failure handlers throws', async () => {
-      numDeletions = 1;
-      const permitterHandlerMock = jest.fn(() =>
-        Promise.reject(new Error('failed')),
-      );
-      const failureHandlerMock = jest.fn(() =>
-        Promise.reject(new Error('failed')),
-      );
-
-      const approvalPromise = approvalController.add({
-        id: 'foo',
-        origin: 'bar.baz',
-        type: 'myType',
-        sideEffects: {
-          permittedHandlers: [permitterHandlerMock],
-          failureHandlers: [failureHandlerMock, failureHandlerMock],
-        },
-      });
-      await approvalController.accept('foo', 'success');
-
-      await expect(async () => approvalPromise).rejects.toThrow(
-        'Unexpected error in side-effects',
-      );
-      expect(permitterHandlerMock).toHaveBeenCalled();
-      expect(failureHandlerMock).toHaveBeenCalledTimes(2);
-      expect(deleteSpy.callCount).toStrictEqual(numDeletions);
-    });
-
-    it('throws on unknown id', async () => {
-      await expect(async () =>
-        approvalController.accept('foo'),
-      ).rejects.toThrow(getIdNotFoundError('foo'));
       expect(deleteSpy.callCount).toStrictEqual(numDeletions);
     });
   });
